# RutasColombiaBack/api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Hello World App
# during dev, should point to your local machine
host: localhost:10010
# basePath prefixes all resource paths 
basePath: /
# 
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
  /hello:
    # binds a127 app logic to a route
    x-swagger-router-controller: hello_world
    get:
      description: Returns 'Hello' to the caller
      # used as the method name of the controller
      operationId: hello
      parameters:
        - name: name
          in: query
          description: The name of the person to whom to say hello
          required: false
          type: string
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/HelloWorldResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /listarMunicipios:
      x-swagger-router-controller: municipios.controlador
      post:
        description: Devuelve todos los Municipios
        operationId: listarMunicipios
        parameters:
          - in: body
            name: body
            required: true
            schema:
              $ref: "#/definitions/Object"
        responses:
          "200":
            description: Success
            schema:
              # a pointer to a definition
              $ref: "#/definitions/Object"
          # responses may fall through to errors
          default:
            description: Error
            schema:
              $ref: "#/definitions/ErrorResponse"             
  /sitiosCercanos:
    x-swagger-router-controller: sitios.controlador
    post:
      description: Devuelve los sitios cercanos a una ubicación geográfica
      operationId: sitiosCercanos
      parameters:
        - in: body
          name: body
          required: true
          schema:
            $ref: "#/definitions/Object"
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/Object"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /listarSitios:
    x-swagger-router-controller: sitios.controlador
    post:
      description: Devuelve todos los sitios
      operationId: listarSitios
      parameters:
        - in: body
          name: body
          required: true
          schema:
            $ref: "#/definitions/Object"
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/Object"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"     
  /crearSitio:
    x-swagger-router-controller: sitios.controlador
    post:
      description: Creación de un sitio
      operationId: crearSitio
      parameters:
        - in: body
          name: body
          required: true
          schema:
            $ref: "#/definitions/Object"
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/Object"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"                 
  /editarSitio:
    x-swagger-router-controller: sitios.controlador
    post:
      description: Edición de un sitio
      operationId: editarSitio
      parameters:
        - in: body
          name: body
          required: true
          schema:
            $ref: "#/definitions/Object"
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/Object"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"          
  /listarPersonas:
    x-swagger-router-controller: personas.controlador
    post:
      description: Devuelve todos los personas
      operationId: listarPersonas
      parameters:
        - in: body
          name: body
          required: true
          schema:
            $ref: "#/definitions/Object"
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/Object"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"     
  /crearPersona:
    x-swagger-router-controller: personas.controlador
    post:
      description: Creación de un persona
      operationId: crearPersona
      parameters:
        - in: body
          name: body
          required: true
          schema:
            $ref: "#/definitions/Object"
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/Object"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"                 
  /editarPersona:
    x-swagger-router-controller: personas.controlador
    post:
      description: Edición de un persona
      operationId: editarPersona
      parameters:
        - in: body
          name: body
          required: true
          schema:
            $ref: "#/definitions/Object"
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/Object"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"        
  /listarAdministradores:
    x-swagger-router-controller: administradores.controlador
    post:
      description: Devuelve todos los administradores
      operationId: listarAdministradores
      parameters:
        - in: body
          name: body
          required: true
          schema:
            $ref: "#/definitions/Object"
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/Object"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"     
  /crearAdministrador:
    x-swagger-router-controller: administradores.controlador
    post:
      description: Creación de un administrador
      operationId: crearAdministrador
      parameters:
        - in: body
          name: body
          required: true
          schema:
            $ref: "#/definitions/Object"
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/Object"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"                 
  /editarAdministrador:
    x-swagger-router-controller: administradores.controlador
    post:
      description: Edición de un administrador
      operationId: editarAdministrador
      parameters:
        - in: body
          name: body
          required: true
          schema:
            $ref: "#/definitions/Object"
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/Object"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"                    
  /swagger:
    x-swagger-pipe: swagger_raw



# complex objects have schema definitions
definitions:
  HelloWorldResponse:
    required:
      - message
    properties:
      message:
        type: string
  
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string

  Object:
    type: object
    
  